another attempt at Dmacos build fixes

+++ runtime/lowlevel_dma.cc
@@ -1101,8 +1101,9 @@ namespace LegionRuntime {
 	for(int i = 0; i < elem_count; i++) {
 	  src_acc.read_untyped(ptr_t(src_index + i), buffer, bytes, src_offset);
           if(0 && i == 0) {
-            printf("remote write: (%ld:%ld->%ld:%ld) %d bytes:",
-                   src_index + i, src_offset, dst_index + i, dst_offset, bytes);
+            printf("remote write: (%zd:%zd->%zd:%zd) %d bytes:",
+                   (ssize_t)(src_index + i), (ssize_t)src_offset,
+                   (ssize_t)(dst_index + i), (ssize_t)dst_offset, bytes);
             for(unsigned j = 0; j < bytes; j++)
               printf(" %02x", (unsigned char)(buffer[j]));
             printf("\n");
